pyproject.toml
[tool.poetry]
name = "mttrchrdsapi"
version = "0.1.0"
description = ""
authors = ["Matt Richards <me@mttrchrds.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
dj-database-url = "^1.3.0"
psycopg2-binary = "^2.9.6"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
pyyaml = "^6.0"
uritemplate = "^4.1.1"
pillow = "^9.5.0"
django-storages = "^1.13.2"
boto3 = "^1.26.114"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
